+++ packages/api/src/index.ts
@@ -1,8 +1,8 @@
 import type { inferRouterInputs, inferRouterOutputs } from "@trpc/server";
 
-import type { AppRouter } from "./trpc/router";
+import type { AppRouter } from "./root";
+import { appRouter } from "./root";
 import { createCallerFactory, createTRPCContext } from "./trpc";
-import { appRouter } from "./trpc/router";
 
 /**
  * Create a server-side caller for the tRPC API
@@ -31,4 +31,3 @@ type RouterOutputs = inferRouterOutputs<AppRouter>;
 
 export { createTRPCContext, appRouter, createCaller };
 export type { AppRouter, RouterInputs, RouterOutputs };
-export { GET, POST } from "./app";
