--- packages/api/src/middlewares/validate-session.ts
@@ -1,28 +1,0 @@
-import type { Context, Next } from "hono";
-import { getCookie } from "hono/cookie";
-
-import { lucia } from "@feprep/auth";
-
-export async function validateSession(c: Context, next: Next) {
-  const sessionId = getCookie(c, lucia.sessionCookieName) ?? null;
-  if (!sessionId) {
-    c.set("user", null);
-    c.set("session", null);
-    return next();
-  }
-  const { session, user } = await lucia.validateSession(sessionId);
-  if (session?.fresh) {
-    // use `header()` instead of `setCookie()` to avoid TS errors
-    c.header("Set-Cookie", lucia.createSessionCookie(session.id).serialize(), {
-      append: true,
-    });
-  }
-  if (!session) {
-    c.header("Set-Cookie", lucia.createBlankSessionCookie().serialize(), {
-      append: true,
-    });
-  }
-  c.set("user", user);
-  c.set("session", session);
-  return next();
-}
