+++ packages/api/src/root.ts
@@ -1,6 +1,6 @@
-import { createTRPCRouter, publicProcedure } from "..";
-import { authRouter } from "./auth";
-import { usersRouter } from "./users";
+import { authRouter } from "./router/auth";
+import { usersRouter } from "./router/users";
+import { createTRPCRouter, publicProcedure } from "./trpc";
 
 export const appRouter = createTRPCRouter({
   hello: publicProcedure.query(() => "hewwo world"),
